Add languages, dynamic image loading to GuideHero
